move mqf2 distribution, fix issues
