+++ src/gluonts/torch/distributions/__init__.py
@@ -28,7 +28,6 @@ from .implicit_quantile_network import (
     ImplicitQuantileNetworkOutput,
 )
 from .isqf import ISQF, ISQFOutput
-from .mqf2 import MQF2Distribution, MQF2DistributionOutput
 from .negative_binomial import NegativeBinomialOutput
 from .piecewise_linear import PiecewiseLinear, PiecewiseLinearOutput
 from .spliced_binned_pareto import (
@@ -53,8 +52,6 @@ __all__ = [
     "ISQF",
     "ISQFOutput",
     "LaplaceOutput",
-    "MQF2Distribution",
-    "MQF2DistributionOutput",
     "NegativeBinomialOutput",
     "NormalOutput",
     "PiecewiseLinear",
